v2.2.75+121
updated club rewards distribution sheet
Updated referral down line count display and show down line referrals
Added usd value for rewards beside the reward amount
Updated logo

+++ lib/utils/number_util.dart
@@ -25,15 +25,17 @@ class NumberUtil {
   /// Breaks at precision 19
   static Decimal decimalLimiter(Decimal input, {int decimalPrecision = 2}) {
     var finalRes = Constants.decimalZero;
-    input ??= Constants.decimalZero;
-    if (input != Constants.decimalZero) {
-      var p = pow(10, decimalPrecision);
-      var t = Decimal.fromInt(p);
-      var x = input * t;
-      var trunc = x.truncate();
-      var resInRational = trunc / t;
-      finalRes =
-          resInRational.toDecimal(scaleOnInfinitePrecision: decimalPrecision);
+    if (input != null) {
+      input ??= Constants.decimalZero;
+      if (input != Constants.decimalZero) {
+        var p = pow(10, decimalPrecision);
+        var t = Decimal.fromInt(p);
+        var x = input * t;
+        var trunc = x.truncate();
+        var resInRational = trunc / t;
+        finalRes =
+            resInRational.toDecimal(scaleOnInfinitePrecision: decimalPrecision);
+      }
     }
     //debugPrint('finalRes $finalRes');
     return finalRes;
